+++ jogadores/JogadorComum.java
@@ -11,8 +11,8 @@ public class JogadorComum extends Jogador {
 
 	@Override
 	protected int movimento() {
-		byte d1 = Dados.lancarDado1();
-		byte d2 = Dados.lancarDado2();
+		int d1 = Dados.lancarDado1();
+		int d2 = Dados.lancarDado2();
 		return d1 + d2;
 	}
